refactor: move `ReversedProds` to `grammar_tools`
