+++ project/cfpq.py
@@ -1,31 +1,13 @@
 from collections import defaultdict
 from dataclasses import dataclass
 from itertools import product
-from typing import Iterable, Optional, cast
+from typing import Optional, cast
 
 import networkx as nx
-from pyformlang.cfg import CFG, Production, Terminal, Variable
+from pyformlang.cfg import CFG, Terminal, Variable
 from scipy.sparse import csr_array
 
-from project.grammar_tools import cfg_to_weak_normal_form
-
-
-class ReversedProds:
-    def __init__(self, productions: Iterable[Production]):
-        self.terminals: dict[Terminal, set[Variable]] = defaultdict(set)
-        self.bodies: dict[tuple[Variable, Variable], set[Variable]] = defaultdict(set)
-
-        for prod in productions:
-            if len(prod.body) == 1 and isinstance(term := prod.body[0], Terminal):
-                self.terminals[term].add(prod.head)
-                continue
-
-            if (
-                len(prod.body) == 2
-                and isinstance(var1 := prod.body[0], Variable)
-                and isinstance(var2 := prod.body[1], Variable)
-            ):
-                self.bodies[(var1, var2)].add(prod.head)
+from project.grammar_tools import ReversedProds, cfg_to_weak_normal_form
 
 
 def hellings_based_cfpq(
