+++ project/grammar_tools.py
@@ -1,6 +1,7 @@
-from typing import cast
+from collections import defaultdict
+from typing import Iterable, cast
 
-from pyformlang.cfg import CFG, Epsilon, Production, Variable
+from pyformlang.cfg import CFG, Epsilon, Production, Terminal, Variable
 
 
 def cfg_to_weak_normal_form(cfg: CFG) -> CFG:
@@ -18,3 +19,21 @@ def cfg_to_weak_normal_form(cfg: CFG) -> CFG:
         start_symbol=nf_cfg.start_symbol,
         productions=new_productions,
     )
+
+
+class ReversedProds:
+    def __init__(self, productions: Iterable[Production]):
+        self.terminals: dict[Terminal, set[Variable]] = defaultdict(set)
+        self.bodies: dict[tuple[Variable, Variable], set[Variable]] = defaultdict(set)
+
+        for prod in productions:
+            if len(prod.body) == 1 and isinstance(term := prod.body[0], Terminal):
+                self.terminals[term].add(prod.head)
+                continue
+
+            if (
+                len(prod.body) == 2
+                and isinstance(var1 := prod.body[0], Variable)
+                and isinstance(var2 := prod.body[1], Variable)
+            ):
+                self.bodies[(var1, var2)].add(prod.head)
